Split tests into modules & fixtures
